Introduced Daemon interface
