+++ packages/devtools/cli/src/base-command.ts
@@ -17,7 +17,7 @@ import * as Sentry from '@dxos/sentry';
 import { captureException } from '@dxos/sentry';
 import * as Telemetry from '@dxos/telemetry';
 
-import { Daemon } from './daemon';
+import { Daemon, Pm2Daemon } from './daemon';
 import {
   disableTelemetry,
   getTelemetryContext,
@@ -230,8 +230,8 @@ export abstract class BaseCommand extends Command {
 
   async execWithDaemon<T>(callback: (daemon: Daemon) => Promise<T | undefined>): Promise<T | undefined> {
     try {
-      const daemon = new Daemon();
-      await daemon.init();
+      const daemon = new Pm2Daemon();
+      await daemon.connect();
       const value = await callback(daemon);
       log('Disconnecting...');
       daemon.disconnect();
