+++ packages/devtools/cli/src/daemon/index.ts
@@ -3,4 +3,5 @@
 //
 
 export * from './daemon';
+export * from './pm2';
 export * from './run';
