+++ packages/devtools/cli/src/daemon/pm2/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export * from './pm2-daemon';
